feat: splits code
