+++ client/src/components/Header/Header.types.ts
@@ -1,0 +1,6 @@
+import { PageNames, PageRoutes } from '../Layout/Layout.types';
+
+export type HeaderProps = {
+  page: PageRoutes;
+  pageNames: PageNames;
+};
