+++ client/src/components/Layout/pageNames.ts
@@ -1,0 +1,12 @@
+import { PageNames } from './Layout.types';
+
+export const pageNames: PageNames = {
+  '/': 'Login',
+  '/planning': 'Planning',
+  '/consultation': 'Consultation',
+  '/dossiers': 'Liste des dossiers patient',
+  '/dossier:id': 'Dossier patient',
+  '/patient/:patientId/dossier': 'Dossier patient',
+  '/admin': 'Administrateur',
+  '/consultations': 'Liste consultations'
+};
